quiz frontend project.
A quiz is an effective aid to the learning process. As quizzes are supposed to test your knowledge, they require considerable time and effort to create. The good news is there are apps available for Android that let you conveniently create and share quizzes. These apps offer many useful options to aid you in the process.